MindGuard/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0
tqdm>=4.65.0
pyyaml>=6.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
onnx>=1.15.0
onnxruntime>=1.15.0
fastapi>=0.100.0
uvicorn>=0.34.0
python-multipart>=0.0.6
